Fix: MissingMethodException in JwtMiddleware by using interface
This commit address a System.MissingMethodException that was being thrown when trying to create an instance of 'Ezenity.Infrastructure.Helpers.AppSettings'.

the issue occured becasue the middleware was expecting a concrete implementation of AppSettings, which didn't have a parameterless constructor.

The solution was to change the type of '_appSettings' from 'AppSettings' to 'IAppSettings' to allow dependency injection to work properly, ensurng that the middleware uses the registered isntance from the DI container.